add support for using AGN templates
add the ability to fit an AGN template created from a previous run of  BADASS. When fitting IFU data of Sy1 galaxies, this makes things a lot less expensive by reducing the number of free parameters to just 1 (the amplitude of the template). No additional fitting for the parameters of broad emission lines, Fe II templates, or the power law.